% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift_matrix.R
\name{plot_drift_matrix}
\alias{plot_drift_matrix}
\title{Plot simulated drift matrix}
\usage{
plot_drift_matrix(freq_matrix)
}
\arguments{
\item{freq_matrix}{Output from `simulate_drift_matrix()`}
}
\value{
A ggplot object
}
\description{
Plot simulated drift matrix
}
