% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift_sample.R
\name{plot_drift_sample}
\alias{plot_drift_sample}
\title{Plot drift simulation of sampled alleles}
\usage{
plot_drift_sample(freq_matrix)
}
\arguments{
\item{freq_matrix}{Output from `simulate_drift_sample()`}
}
\value{
A base R plot
}
\description{
Plot drift simulation of sampled alleles
}
