% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_drift_matrix.R
\name{simulate_drift_matrix}
\alias{simulate_drift_matrix}
\title{Simulate allele frequency changes due to genetic drift (Wright–Fisher model)}
\usage{
simulate_drift_matrix(N, p, G, nsim)
}
\arguments{
\item{N}{Number of diploid individuals}

\item{p}{Initial frequency of allele A1}

\item{G}{Number of generations}

\item{nsim}{Number of simulation replicates}
}
\value{
A data frame of allele frequencies across generations and simulations
}
\description{
Simulate allele frequency changes due to genetic drift (Wright–Fisher model)
}
